requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
diffusers==0.27.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2024.3.0
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
hyperframe==6.0.1
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
portalocker==2.8.2
protobuf==4.25.3
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pypdfium2==4.28.0
python-dotenv==1.0.1
PyYAML==6.0.1
qdrant-client==1.8.0
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sniffio==1.3.1
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
triton==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
uuid==1.30
Werkzeug==3.0.1
zipp==3.18.1
